backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.1
django-cors-headers>=4.4
dj-database-url>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
uvicorn>=0.27
whitenoise>=6.7
